Split chips.rs, to chips.rs and generator.rs, fixed some clippy warnings
* `chips.rs` is responsible for parsing, similar naming scheme as e.g,
  `dma.rs`, `interrupts.rs`, etc.
* `generator.rs` contains the logic and JSON emitting code.
